implement About and Content lazy-loading
