+++ src/features/Content/ContentEditor.js
@@ -1,6 +1,4 @@
 import React, { useMemo, useState } from 'react';
-import styled from 'styled-components/macro';
-import { motion } from 'framer-motion';
 
 // prismjs theme
 import 'assets/css/prism-atom-dark.css';
@@ -13,9 +11,8 @@ import Editable from './Editable'
 import ContentFooter from 'features/ContentFooter/ContentFooter';
 
 
-const ContentEditor = ({ content, updateContent, shouldDelayAnimation }) => {
-    // must create a new instance whenever a new content item is selected
-    const editor = useMemo(() => createEditor(), [content.id])
+const ContentEditor = ({ content, updateContent }) => {
+    const editor = useMemo(() => createEditor(), [])
 
     const [ editorState, setEditorState ] = useState(content.data);
     const [ readOnly, setReadOnly ] = useState(true);
@@ -27,16 +24,7 @@ const ContentEditor = ({ content, updateContent, shouldDelayAnimation }) => {
     })
 
     return (
-        // variants and transition definitions are below
-        <Wrapper variants={wrapper}
-                initial='exit'
-                animate='enter'
-                exit='exit'
-                transition={shouldDelayAnimation ? {
-                    ...enterTransition,
-                    delay: .85,
-                } : enterTransition }>
-
+        <>
             <ToggleReadOnly readOnly={readOnly}
                 toggle={() => {
                     // must nullify the selection when quitting editing
@@ -44,8 +32,7 @@ const ContentEditor = ({ content, updateContent, shouldDelayAnimation }) => {
                     // when the user will start editing again
                     if (!readOnly) editor.selection = null
                     setReadOnly(!readOnly)
-                }} 
-            />
+                }} />
 
             <Slate editor={editor} 
                 value={editorState} 
@@ -58,43 +45,8 @@ const ContentEditor = ({ content, updateContent, shouldDelayAnimation }) => {
                     edited={content.edited}
                     saveChanges={saveChanges} />
             </Slate>
-        </Wrapper>
+        </>
     )
-}
-
-
-const Wrapper = styled(motion.div)`
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    margin: 6px 9px;
-    height: calc(100% - 12px);
-    font-size: 1.4rem;
-    color: var(--gray1); 
-`;
-
-
-const enterTransition = {
-    duration: 0.3,
-    when: 'beforeChildren',
-    staggerChildren: 0.09,
-    ease: 'circOut'
-}
-
-
-const wrapper = {
-    enter: {
-        scale: 1,
-        opacity: 1,
-    },
-    exit: {
-        scale: 0.7,
-        opacity: 0,
-        transition: {
-            duration: 0.4,
-            ease: 'circOut'
-        }
-    }
 }
 
 export default ContentEditor;
